Move SeasonsRow to a separate view.
